+++ src/components/auth/Login/index.tsx
@@ -16,7 +16,7 @@ import { inputId } from 'common/constant/input';
 import { Input, Button } from 'atoms';
 import routes from 'common/constant/routes';
 import Router from 'next/router';
-import GoogleLoginButton from 'components/ui/GoogleLoginButton';
+import GoogleLoginButton from 'components/common/GoogleLoginButton';
 import * as S from '../style';
 
 const Login = (): React.ReactElement => {
